WE:AD changes in playlist. Read description.
Including every map in the playlists is a good idea.
The present issue is WE:AD not being shipped in retail season 3.
It's a separate download.

Setting it to 1 will include it in the randomized map select.
Setting it to 0 will make it able to be used with the playlist but will be exluded from said randomized map select.

It's set to 0 now to avoid the game from closing incase it hits said scenario.