Fix CVE-2022-27774
Merge: curl/curl@139a54ed0a172ada
Note: protocol compare couldn't be added, as this would require adding a new member in the state structure, and therefore, break compatibility with the structures in the compiled executable (breaking the ability to hook any of its code). This fix should be sufficient however.